Fixed variable substitution in pipeline.
