+++ src/pipeline/__init__.py
@@ -1,6 +1,8 @@
 from .varhelper import *
+from .pipelinetreenode import *
 from .pipelinetree import *
 
 __all__ = []
 __all__.extend(varhelper.__all__)
+__all__.extend(pipelinetreenode.__all__)
 __all__.extend(pipelinetree.__all__)
