+++ src/pipeline/varhelper.py
@@ -80,8 +80,8 @@ def get_referenced_variable_value(line: str, scopes: Dict[str, Dict[str, object]
         if key is not None:
             return key, value
 
-        # print("\n\n")
-        # print(scopes["THIS"])
+        print("\n\n")
+        print(scopes["THIS"])
 
         # Raise exception if not found in current scope
         raise ValueError(__name__ + ": Error '{}' is unknown within the scope '{}'.".format(line, key))
